Actualizados tipos de datos a decimal en lugar de int

+++ TravelioREST/Mesas/MesaCancelarReserva.cs
@@ -19,7 +19,7 @@ namespace TravelioREST.Mesas;
 public class ReservaCancelarResponse
 {
     public bool exito { get; set; }
-    public int valorPagado { get; set; }
+    public decimal valorPagado { get; set; }
 }
 
 public static class MesaCancelarReserva
